Rename components into appropriate folders
