+++ src/components/ExternalNetworkIcon/index.jsx
